Create Question_3
Here everything in the problem statement is covered and 1 thing to specify is that it does not show what incorrect format we have entered (for eg. if i haven't entered a capital letter then it will show "Invalid password format" but will not show what format is missing because it may harm the security.